Create Finding area of simple polygon in O(N)
This is easy to do if we go through all edges and add trapezoid areas bounded by each edge and x-axis. The area needs to be taken with sign so that the extra area will be reduced. Hence, the formula is as follows:

A=∑(p,q)∈edges(px−qx)⋅(py+qy)2