Create Minkowski sum of convex polygons
One of the most common applications of Minkowski sum is computing the distance between two convex polygons (or simply checking whether they intersect). The distance between two convex polygons P and Q is defined as mina∈P,b∈Q||a−b||. One can note that the distance is always attained between two vertices or a vertex and an edge, so we can easily find the distance in O(|P||Q|). However, with clever usage of Minkowski sum we can reduce the complexity to O(|P|+|Q|).

If we reflect Q through the point (0,0) obtaining polygon −Q, the problem boils down to finding the smallest distance between a point in P+(−Q) and (0,0). We can find that distance in linear time using the following idea. If (0,0) is inside or on the boundary of polygon, the distance is 0, otherwise the distance is attained between (0,0) and some vertex or edge of the polygon. Since Minkowski sum can be computed in linear time, we obtain a linear-time algorithm for finding the distance between two convex polygons.

Implementation
Below is the implementation of Minkowski sum for polygons with integer points. Note that in this case all computations can be done in integers since instead of computing polar angles and directly comparing them we can look at the sign of cross product of two vectors.