Create Convex hull trick and Li Chao tree
Consider the following problem. There are n cities. You want to travel from city 1 to city n by car. To do this you have to buy some gasoline. It is known that a liter of gasoline costs costk in the kth city. Initially your fuel tank is empty and you spend one liter of gasoline per kilometer. Cities are located on the same line in ascending order with kth city having coordinate xk. Also you have to pay tollk to enter kth city. Your task is to make the trip with minimum possible cost. It's obvious that the solution can be calculated via dynamic programming:

dpi=tolli+minj<i(costj⋅(xi−xj)+dpj)
Naive approach will give you O(n2) complexity which can be improved to O(nlogn) or O(nlog[Cε−1]) where C is largest possible |xi| and ε is precision with which xi is considered (ε=1 for integers which is usually the case). To do this one should note that the problem can be reduced to adding linear functions k⋅x+b to the set and finding minimum value of the functions in some particular point x. There are two main approaches one can use here.