Added a program on Stack using Linked List
Stack and Linked List are important parts of Data Structures and Algorithms. And when combined together it can be quite confusing!
Here's a simple well structured program on the implementation of Stack ADT using Linked List, together with the addition of a switch case for the user's convenience and ease of understanding.